refactor(logger): groups methods by receiver
